pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "twopidgeons"
version = "0.1.0"
description = "A distributed image management library secured by blockchain"
authors = [
  { name="Your Name", email="your.email@example.com" },
]
dependencies = [
    "Pillow",
]
requires-python = ">=3.8"

[tool.setuptools]
packages = ["twopidgeons"]

[project.scripts]
twopidgeons = "twopidgeons.cli:main"
